Refactor pkgops related pkgs from "root" to "ipc" as it is now shared between shizuku and root.
