+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/pkgs/pkgops/ipc/PkgOpsConnection.aidl
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.pkgs.pkgops.root;
+package eu.darken.sdmse.common.pkgs.pkgops.ipc;
 
 import android.content.pm.PackageInfo;
 import eu.darken.sdmse.common.ipc.RemoteInputStream;
