+++ app-common-io/src/main/java/eu/darken/sdmse/common/pkgs/pkgops/ipc/PkgOpsClient.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.pkgs.pkgops.root
+package eu.darken.sdmse.common.pkgs.pkgops.ipc
 
 import android.content.pm.PackageInfo
 import dagger.assisted.Assisted
