+++ app-common-io/src/main/java/eu/darken/sdmse/common/pkgs/pkgops/ipc/PkgOpsHost.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.pkgs.pkgops.root
+package eu.darken.sdmse.common.pkgs.pkgops.ipc
 
 import android.app.ActivityManager
 import android.content.Context
